+++ src/@ionic-native/plugins/pay-pal/index.ts
@@ -227,7 +227,7 @@ export class PayPalItem {
    * @param {String} currency: ISO standard currency code.
    * @param {String} sku: The stock keeping unit for this item. 50 characters max (optional)
    */
-  constructor(name: string, quantity: number, price: string, currency: string, sku: string) {
+  constructor(name: string, quantity: number, price: string, currency: string, sku?: string) {
     this.name = name;
     this.quantity = quantity;
     this.price = price;
@@ -253,7 +253,7 @@ export class PayPalItem {
   /**
    * The stock keeping unit for this item. 50 characters max (optional)
    */
-  sku: string;
+  sku?: string;
 }
 
 /**